requirements/development.txt
coverage==4.1
flake8==2.5.4
ipython==4.2.0
mixer==5.5.7
pytest-cov==2.2.1
pytest-django==2.9.1
pytest-mock==1.0
pytest-pythonpath==0.7
pytest==2.9.1
redislite==3.0.296
